+++ packages/shared/src/types/userFeedSubscriptions.types.ts
@@ -1,26 +1,14 @@
-import {z} from 'zod';
-
 import type {AccountId} from '@shared/types/accounts.types';
 import type {DeliverySchedule} from '@shared/types/deliverySchedules.types';
-import type {PersistedFeedSourceType} from '@shared/types/feedSourceTypes.types';
-import {FeedSourceType, PERSISTED_FEED_SOURCE_TYPES} from '@shared/types/feedSourceTypes.types';
-import {FirestoreTimestampSchema} from '@shared/types/firebase.types';
+import type {FeedSourceType, PersistedFeedSourceType} from '@shared/types/feedSourceTypes.types';
 import type {BaseStoreItem} from '@shared/types/utils.types';
-import {YouTubeChannelIdSchema, type YouTubeChannelId} from '@shared/types/youtube.types';
-
-import {AccountIdSchema} from '@shared/schemas/accounts.schema';
-import {DeliveryScheduleFromStorageSchema} from '@shared/schemas/deliverySchedules.schema';
+import type {YouTubeChannelId} from '@shared/types/youtube.types';
 
 /**
  * Strongly-typed type for a {@link UserFeedSubscription}'s unique identifier. Prefer this over
  * plain strings.
  */
 export type UserFeedSubscriptionId = string & {readonly __brand: 'UserFeedSubscriptionIdBrand'};
-
-/**
- * Zod schema for a {@link UserFeedSubscriptionId}.
- */
-export const UserFeedSubscriptionIdSchema = z.string().uuid();
 
 /**
  * An individual account's subscription to a feed source.
@@ -66,47 +54,3 @@ export type UserFeedSubscription =
   | RssUserFeedSubscription
   | YouTubeChannelUserFeedSubscription
   | IntervalUserFeedSubscription;
-
-/**
- * Zod schema for a {@link UserFeedSubscription} persisted to Firestore.
- */
-const BaseUserFeedSubscriptionFromStorageSchema = z.object({
-  feedSourceType: z.enum(PERSISTED_FEED_SOURCE_TYPES),
-  userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
-  accountId: AccountIdSchema,
-  isActive: z.boolean(),
-  deliverySchedule: DeliveryScheduleFromStorageSchema,
-  unsubscribedTime: FirestoreTimestampSchema.or(z.date()).optional(),
-  createdTime: FirestoreTimestampSchema.or(z.date()),
-  lastUpdatedTime: FirestoreTimestampSchema.or(z.date()),
-});
-
-export const RssUserFeedSubscriptionFromStorageSchema =
-  BaseUserFeedSubscriptionFromStorageSchema.extend({
-    feedSourceType: z.literal(FeedSourceType.RSS),
-    url: z.string().url(),
-    title: z.string(),
-  });
-
-export const YouTubeChannelUserFeedSubscriptionFromStorageSchema =
-  BaseUserFeedSubscriptionFromStorageSchema.extend({
-    feedSourceType: z.literal(FeedSourceType.YouTubeChannel),
-    channelId: YouTubeChannelIdSchema,
-  });
-
-export const IntervalUserFeedSubscriptionFromStorageSchema =
-  BaseUserFeedSubscriptionFromStorageSchema.extend({
-    feedSourceType: z.literal(FeedSourceType.Interval),
-    intervalSeconds: z.number().positive().int().min(60),
-  });
-
-export const UserFeedSubscriptionFromStorageSchema = z.union([
-  RssUserFeedSubscriptionFromStorageSchema,
-  YouTubeChannelUserFeedSubscriptionFromStorageSchema,
-  IntervalUserFeedSubscriptionFromStorageSchema,
-]);
-
-/**
- * Type for a {@link UserFeedSubscription} persisted to Firestore.
- */
-export type UserFeedSubscriptionFromStorage = z.infer<typeof UserFeedSubscriptionFromStorageSchema>;
